feat: add mutation stock report exports and enhance mutation handling
- Added MutationStockReportExport and MutationStockExport classes for Excel and PDF exports of mutation stock reports.
- Created new views for mutation stock report exports in both Excel (mutation_stock_report.blade.php) and PDF (mutation_stock_report_pdf.blade.php) formats.
- Extended `api.php` routes to include new export endpoints for mutation stock.
- Updated MutationController to support report generating and exporting functionality.
- Refactored MutationRepo to improve mutation handling and add support for status updates (open, parsial, close) for mutation_out records.

+++ src/Repositories/Pembelian/Invoice/InvoiceRepo.php
@@ -699,7 +699,7 @@ class InvoiceRepo extends ElequentRepository
             if(!empty($findInvoice) && $findInvoice->invoice_status == StatusEnum::BELUM_LUNAS){
                 PurchasePaymentInvoice::create([
                     'invoice_no' => $findInvoice->invoice_no,
-                    'total_payment' => Utility::remove_commas($total),
+                    'total_payment' => (float) Utility::remove_commas($total),
                     'payment_date' => $returDate,
                     'total_discount' => '0',
                     'coa_id_discount' => "",
